refactor folder structure step 2
